+++ docs/assets/Theming.stories-LH0q8yvu.js
@@ -1,4 +1,4 @@
-import{D as r,a as j,j as e,T as a,H as t,W as d,b as s}from"./constants-N744Lllm.js";import"./index-4g5l5LRQ.js";import"./_commonjsHelpers-4gQjN7DL.js";const T={title:"Example/Theming",component:r,parameters:{layout:"fullscreen",deepControls:{enabled:!0}},argTypes:j},l={render:n=>e.jsx("div",{style:{background:"#fff",boxShadow:"0 2px 5px rgba(0, 0, 0, 0.1)",width:"300px",borderRadius:"0px",margin:"30px auto",padding:"10px"},children:e.jsxs(r,{isRange:!1,initialValue:new Date,calendar:"gregory",config:{weekStartsOn:"monday",weekdayFormat:"short",weekends:["saturday","sunday"]},...n,onChange:o=>{console.log("onChange: ",o)},children:[e.jsx(a,{style:{color:"#443f70",fontWeight:"300"}}),e.jsx(t,{monthSelectStyles:{backgroundColor:"#f0f0f0",color:"#000",padding:"4px",borderRadius:"5px",outline:"none"},yearSelectStyles:{backgroundColor:"#f0f0f0",color:"#000",padding:"4px",borderRadius:"5px",outline:"none"}}),e.jsx(d,{style:{fontWeight:400}}),e.jsx(s,{slotStyles:{borderRadius:"50%",fontWeight:"300"},selectedStyles:{borderRadius:"50%",backgroundColor:"#783aeb",fontWeight:"500",color:"#fff"},selectedParentStyles:{borderRadius:"50%"},weekendStyles:{fontStyle:"italic"}})]})})},i={render:n=>e.jsx("div",{style:{background:"#fff",boxShadow:"0 2px 5px rgba(0, 0, 0, 0.1)",width:"350px",borderRadius:"0px",margin:"30px auto",padding:"10px"},children:e.jsxs(r,{isRange:!0,initialValue:[new Date("2024-05-15"),new Date("2024-05-18")],calendar:"gregory",config:{weekStartsOn:"monday",weekdayFormat:"short",weekends:["saturday","sunday"]},...n,onChange:o=>{console.log("onChange: ",o)},children:[e.jsx(a,{}),e.jsx(t,{monthSelectStyles:{backgroundColor:"#f0f0f0",color:"#000",padding:"4px",borderRadius:"5px",outline:"none"},yearSelectStyles:{backgroundColor:"#f0f0f0",color:"#000",padding:"4px",borderRadius:"5px",outline:"none"}}),e.jsx(d,{style:{fontWeight:400}}),e.jsx(s,{slotStyles:{fontWeight:"300"},selectedStyles:{borderRadius:"10px",backgroundColor:"#303030",fontWeight:"500"},startOfRangeParentStyles:{borderTopLeftRadius:"10px",borderBottomLeftRadius:"10px"},endOfRangeParentStyles:{borderTopRightRadius:"10px",borderBottomRightRadius:"10px"},inSelectedRangeParentStyles:{backgroundColor:"#a8adb4"},inSelectedRangeStyles:{color:"#fff"},slotParentStyles:{padding:"3px"}})]})})},g={render:n=>e.jsx("div",{style:{background:"#fff",boxShadow:"0 2px 5px rgba(0, 0, 0, 0.1)",width:"300px",borderRadius:"0px",margin:"30px auto",padding:"10px",border:"1px solid #acb5be"},children:e.jsxs(r,{isRange:!0,initialValue:[new Date("2024-05-15"),new Date("2024-05-18")],calendar:"gregory",config:{weekStartsOn:"monday",weekdayFormat:"short",weekends:["saturday","sunday"]},...n,onChange:o=>{console.log("onChange: ",o)},children:[e.jsx(a,{}),e.jsx(t,{monthSelectStyles:{backgroundColor:"#f0f0f0",color:"#000",padding:"4px",borderRadius:"5px",outline:"none"},yearSelectStyles:{backgroundColor:"#f0f0f0",color:"#000",padding:"4px",borderRadius:"5px",outline:"none"}}),e.jsx(d,{style:{fontWeight:300,marginBottom:"4px",color:"#7a7a7a"}}),e.jsx(s,{parentStyles:{gap:"2px"},slotParentStyles:{border:"1px solid #c7cbcf"},slotStyles:{fontWeight:"300",borderRadius:0},selectedStyles:{backgroundColor:"#d80202",fontWeight:"500"},startOfRangeParentStyles:{borderRadius:0},endOfRangeParentStyles:{borderRadius:0},inSelectedRangeParentStyles:{backgroundColor:"#ff9a9a"},inSelectedRangeStyles:{color:"#fff"}})]})})},c={render:n=>e.jsx("div",{style:{background:"#333333",boxShadow:"0 2px 5px rgba(0, 0, 0, 0.1)",width:"360px",borderRadius:"12px",margin:"30px auto",padding:"17px"},children:e.jsxs(r,{isRange:!0,initialValue:[new Date("2024-05-15"),new Date("2024-05-18")],calendar:"gregory",config:{weekStartsOn:"monday",weekdayFormat:"short",weekends:["saturday","sunday"]},...n,onChange:o=>{console.log("onChange: ",o)},children:[e.jsx(a,{style:{color:"#fff"}}),e.jsx(t,{monthSelectStyles:{backgroundColor:"#636363",color:"#fff",padding:"4px",borderRadius:"5px",outline:"none"},yearSelectStyles:{backgroundColor:"#636363",color:"#fff",padding:"4px",borderRadius:"5px",outline:"none"}}),e.jsx(d,{style:{fontWeight:300,marginBottom:"4px",color:"#7a7a7a"}}),e.jsx(s,{slotStyles:{color:"#fff",fontWeight:"300"},slotParentStyles:{border:"0",padding:"4px"},weekendStyles:{color:"#ff4794"},selectedStyles:{backgroundColor:"#ff4794",fontWeight:"500"},endOfRangeStyles:{color:"#ffffff"},startOfRangeStyles:{color:"#ffffff"},inSelectedRangeParentStyles:{backgroundColor:"#6b6b6b"}})]})})},f={render:n=>e.jsx("div",{style:{background:"#fff",boxShadow:"0 2px 5px rgba(0, 0, 0, 0.1)",width:"300px",borderRadius:"0px",margin:"30px auto",padding:"10px",border:"1px solid #acb5be"},children:e.jsxs(r,{isRange:!0,initialValue:[new Date("2024-05-15"),new Date("2024-05-18")],calendar:"gregory",config:{weekStartsOn:"monday",weekdayFormat:"short",weekends:["saturday","sunday"],showOtherDays:!0},...n,onChange:o=>{console.log("onChange: ",o)},children:[e.jsx(a,{style:{fontWeight:400}}),e.jsx(t,{monthSelectStyles:{backgroundColor:"#f0f0f0",color:"#000",padding:"4px",borderRadius:"5px",outline:"none"},yearSelectStyles:{backgroundColor:"#f0f0f0",color:"#000",padding:"4px",borderRadius:"5px",outline:"none"}}),e.jsx(d,{style:{fontWeight:300,marginBottom:"4px",color:"#7a7a7a"}}),e.jsx(s,{parentStyles:{gap:"1px",backgroundColor:"#d6dadf",border:"1px solid #d6dadf"},slotParentStyles:{backgroundColor:"#fff",border:0},slotStyles:{fontWeight:"300",borderRadius:0,color:"#494949"},selectedStyles:{backgroundColor:"#1dfc00",fontWeight:"500"},weekendStyles:{color:"#ff0000"},startOfRangeParentStyles:{borderRadius:0},endOfRangeParentStyles:{borderRadius:0},inSelectedRangeParentStyles:{backgroundColor:"#d5f8d5"},inSelectedRangeStyles:{color:"#000000"}})]})})};var p,u,y;l.parameters={...l.parameters,docs:{...(p=l.parameters)==null?void 0:p.docs,source:{originalSource:`{
+import{D as r,a as O,j as e,T as a,H as t,W as d,b as s}from"./constants-nuSoQ8zG.js";import"./index-4g5l5LRQ.js";import"./_commonjsHelpers-4gQjN7DL.js";const T={title:"Example/Theming",component:r,parameters:{layout:"fullscreen",deepControls:{enabled:!0}},argTypes:O},l={render:n=>e.jsx("div",{style:{background:"#fff",boxShadow:"0 2px 5px rgba(0, 0, 0, 0.1)",width:"300px",borderRadius:"0px",margin:"30px auto",padding:"10px"},children:e.jsxs(r,{isRange:!1,initialValue:new Date,calendar:"gregory",config:{weekStartsOn:"monday",weekdayFormat:"short",weekends:["saturday","sunday"]},...n,onChange:o=>{console.log("onChange: ",o)},children:[e.jsx(a,{style:{color:"#443f70",fontWeight:"300"}}),e.jsx(t,{monthSelectStyles:{backgroundColor:"#f0f0f0",color:"#000",padding:"4px",borderRadius:"5px",outline:"none"},yearSelectStyles:{backgroundColor:"#f0f0f0",color:"#000",padding:"4px",borderRadius:"5px",outline:"none"}}),e.jsx(d,{style:{fontWeight:400}}),e.jsx(s,{slotStyles:{borderRadius:"50%",fontWeight:"300"},selectedStyles:{borderRadius:"50%",backgroundColor:"#783aeb",fontWeight:"500",color:"#fff"},selectedParentStyles:{borderRadius:"50%"},weekendStyles:{fontStyle:"italic"}})]})})},i={render:n=>e.jsx("div",{style:{background:"#fff",boxShadow:"0 2px 5px rgba(0, 0, 0, 0.1)",width:"350px",borderRadius:"0px",margin:"30px auto",padding:"10px"},children:e.jsxs(r,{isRange:!0,initialValue:[new Date("2024-05-15"),new Date("2024-05-18")],calendar:"gregory",config:{weekStartsOn:"monday",weekdayFormat:"short",weekends:["saturday","sunday"]},...n,onChange:o=>{console.log("onChange: ",o)},children:[e.jsx(a,{}),e.jsx(t,{monthSelectStyles:{backgroundColor:"#f0f0f0",color:"#000",padding:"4px",borderRadius:"5px",outline:"none"},yearSelectStyles:{backgroundColor:"#f0f0f0",color:"#000",padding:"4px",borderRadius:"5px",outline:"none"}}),e.jsx(d,{style:{fontWeight:400}}),e.jsx(s,{slotStyles:{fontWeight:"300"},selectedStyles:{borderRadius:"10px",backgroundColor:"#303030",fontWeight:"500"},startOfRangeParentStyles:{borderTopLeftRadius:"10px",borderBottomLeftRadius:"10px"},endOfRangeParentStyles:{borderTopRightRadius:"10px",borderBottomRightRadius:"10px"},inSelectedRangeParentStyles:{backgroundColor:"#a8adb4"},inSelectedRangeStyles:{color:"#fff"},slotParentStyles:{padding:"3px"}})]})})},c={render:n=>e.jsx("div",{style:{background:"#fff",boxShadow:"0 2px 5px rgba(0, 0, 0, 0.1)",width:"300px",borderRadius:"0px",margin:"30px auto",padding:"10px",border:"1px solid #acb5be"},children:e.jsxs(r,{isRange:!0,initialValue:[new Date("2024-05-15"),new Date("2024-05-18")],calendar:"gregory",config:{weekStartsOn:"monday",weekdayFormat:"short",weekends:["saturday","sunday"]},...n,onChange:o=>{console.log("onChange: ",o)},children:[e.jsx(a,{}),e.jsx(t,{monthSelectStyles:{backgroundColor:"#f0f0f0",color:"#000",padding:"4px",borderRadius:"5px",outline:"none"},yearSelectStyles:{backgroundColor:"#f0f0f0",color:"#000",padding:"4px",borderRadius:"5px",outline:"none"}}),e.jsx(d,{style:{fontWeight:300,marginBottom:"4px",color:"#7a7a7a"}}),e.jsx(s,{parentStyles:{gap:"2px"},slotParentStyles:{border:"1px solid #c7cbcf"},slotStyles:{fontWeight:"300",borderRadius:0},selectedStyles:{backgroundColor:"#d80202",fontWeight:"500"},startOfRangeParentStyles:{borderRadius:0},endOfRangeParentStyles:{borderRadius:0},inSelectedRangeParentStyles:{backgroundColor:"#ff9a9a"},inSelectedRangeStyles:{color:"#fff"}})]})})},g={render:n=>e.jsx("div",{style:{background:"#333333",boxShadow:"0 2px 5px rgba(0, 0, 0, 0.1)",width:"360px",borderRadius:"12px",margin:"30px auto",padding:"17px"},children:e.jsxs(r,{isRange:!0,initialValue:[new Date("2024-05-15"),new Date("2024-05-18")],calendar:"gregory",config:{weekStartsOn:"monday",weekdayFormat:"short",weekends:["saturday","sunday"]},...n,onChange:o=>{console.log("onChange: ",o)},children:[e.jsx(a,{style:{color:"#fff"}}),e.jsx(t,{monthSelectStyles:{backgroundColor:"#636363",color:"#fff",padding:"4px",borderRadius:"5px",outline:"none"},yearSelectStyles:{backgroundColor:"#636363",color:"#fff",padding:"4px",borderRadius:"5px",outline:"none"},prevButtonStyles:{color:"#fff"},nextButtonStyles:{color:"#fff"}}),e.jsx(d,{style:{fontWeight:300,marginBottom:"4px",color:"#7a7a7a"}}),e.jsx(s,{slotStyles:{color:"#fff",fontWeight:"300"},slotParentStyles:{border:"0",padding:"4px"},weekendStyles:{color:"#ff4794"},selectedStyles:{backgroundColor:"#ff4794",fontWeight:"500"},endOfRangeStyles:{color:"#ffffff"},startOfRangeStyles:{color:"#ffffff"},inSelectedRangeParentStyles:{backgroundColor:"#6b6b6b"},inHoveredRangeParentStyles:{backgroundColor:"#525252"}})]})})},f={render:n=>e.jsx("div",{style:{background:"#fff",boxShadow:"0 2px 5px rgba(0, 0, 0, 0.1)",width:"300px",borderRadius:"0px",margin:"30px auto",padding:"10px",border:"1px solid #acb5be"},children:e.jsxs(r,{isRange:!0,initialValue:[new Date("2024-05-15"),new Date("2024-05-18")],calendar:"gregory",config:{weekStartsOn:"monday",weekdayFormat:"short",weekends:["saturday","sunday"],showOtherDays:!0},...n,onChange:o=>{console.log("onChange: ",o)},children:[e.jsx(a,{style:{fontWeight:200,fontSize:"1.2em",marginBottom:"10px",textAlign:"center",borderRadius:"7px",letterSpacing:"16px",textTransform:"uppercase",color:"#5c5c5c"}}),e.jsx(t,{monthSelectStyles:{backgroundColor:"#f0f0f0",color:"#000",padding:"4px",borderRadius:"5px",outline:"none"},yearSelectStyles:{backgroundColor:"#f0f0f0",color:"#000",padding:"4px",borderRadius:"5px",outline:"none"},rootStyles:{boxShadow:"0 2px 5px rgba(0, 0, 0, 0.1) inset",borderRadius:"7px",marginBottom:"10px",padding:"10px",backgroundColor:"#fdfdfd"}}),e.jsx(d,{style:{fontWeight:300,marginBottom:"4px",color:"#7a7a7a"}}),e.jsx(s,{parentStyles:{gap:"1px",backgroundColor:"#d6dadf",border:"1px solid #d6dadf"},slotParentStyles:{backgroundColor:"#fff",border:0},slotStyles:{fontWeight:"300",borderRadius:0,color:"#494949"},selectedStyles:{backgroundColor:"#a4fc00",fontWeight:"500"},weekendStyles:{color:"#ff0000"},startOfRangeParentStyles:{borderRadius:0},endOfRangeParentStyles:{borderRadius:0},startOfRangeStyles:{color:"#288f00"},endOfRangeStyles:{color:"#288f00"},inSelectedRangeParentStyles:{backgroundColor:"#e7ffb9"},inHoveredRangeParentStyles:{backgroundColor:"#ecf8d5"},inSelectedRangeStyles:{color:"#000000"}})]})})};var p,u,x;l.parameters={...l.parameters,docs:{...(p=l.parameters)==null?void 0:p.docs,source:{originalSource:`{
   render: props => {
     return <div style={{
       background: "#fff",
@@ -57,7 +57,7 @@ import{D as r,a as j,j as e,T as a,H as t,W as d,b as s}from"./constants-N744Lll
         </DatePickerProvider>
       </div>;
   }
-}`,...(y=(u=l.parameters)==null?void 0:u.docs)==null?void 0:y.source}}};var x,b,S;i.parameters={...i.parameters,docs:{...(x=i.parameters)==null?void 0:x.docs,source:{originalSource:`{
+}`,...(x=(u=l.parameters)==null?void 0:u.docs)==null?void 0:x.source}}};var y,b,S;i.parameters={...i.parameters,docs:{...(y=i.parameters)==null?void 0:y.docs,source:{originalSource:`{
   render: props => {
     return <div style={{
       background: "#fff",
@@ -118,7 +118,7 @@ import{D as r,a as j,j as e,T as a,H as t,W as d,b as s}from"./constants-N744Lll
         </DatePickerProvider>
       </div>;
   }
-}`,...(S=(b=i.parameters)==null?void 0:b.docs)==null?void 0:S.source}}};var h,k,R;g.parameters={...g.parameters,docs:{...(h=g.parameters)==null?void 0:h.docs,source:{originalSource:`{
+}`,...(S=(b=i.parameters)==null?void 0:b.docs)==null?void 0:S.source}}};var h,R,k;c.parameters={...c.parameters,docs:{...(h=c.parameters)==null?void 0:h.docs,source:{originalSource:`{
   render: props => {
     return <div style={{
       background: "#fff",
@@ -186,7 +186,7 @@ import{D as r,a as j,j as e,T as a,H as t,W as d,b as s}from"./constants-N744Lll
         </DatePickerProvider>
       </div>;
   }
-}`,...(R=(k=g.parameters)==null?void 0:k.docs)==null?void 0:R.source}}};var m,w,C;c.parameters={...c.parameters,docs:{...(m=c.parameters)==null?void 0:m.docs,source:{originalSource:`{
+}`,...(k=(R=c.parameters)==null?void 0:R.docs)==null?void 0:k.source}}};var m,w,C;g.parameters={...g.parameters,docs:{...(m=g.parameters)==null?void 0:m.docs,source:{originalSource:`{
   render: props => {
     return <div style={{
       background: "#333333",
@@ -223,6 +223,10 @@ import{D as r,a as j,j as e,T as a,H as t,W as d,b as s}from"./constants-N744Lll
           padding: "4px",
           borderRadius: "5px",
           outline: "none"
+        }} prevButtonStyles={{
+          color: "#fff"
+        }} nextButtonStyles={{
+          color: "#fff"
         }} />
           <WeekDays style={{
           fontWeight: 300,
@@ -246,11 +250,13 @@ import{D as r,a as j,j as e,T as a,H as t,W as d,b as s}from"./constants-N744Lll
           color: "#ffffff"
         }} inSelectedRangeParentStyles={{
           backgroundColor: "#6b6b6b"
+        }} inHoveredRangeParentStyles={{
+          backgroundColor: "#525252"
         }} />
         </DatePickerProvider>
       </div>;
   }
-}`,...(C=(w=c.parameters)==null?void 0:w.docs)==null?void 0:C.source}}};var P,D,W;f.parameters={...f.parameters,docs:{...(P=f.parameters)==null?void 0:P.docs,source:{originalSource:`{
+}`,...(C=(w=g.parameters)==null?void 0:w.docs)==null?void 0:C.source}}};var P,D,W;f.parameters={...f.parameters,docs:{...(P=f.parameters)==null?void 0:P.docs,source:{originalSource:`{
   render: props => {
     return <div style={{
       background: "#fff",
@@ -275,7 +281,14 @@ import{D as r,a as j,j as e,T as a,H as t,W as d,b as s}from"./constants-N744Lll
         console.log("onChange: ", e);
       }}>
           <Title style={{
-          fontWeight: 400
+          fontWeight: 200,
+          fontSize: "1.2em",
+          marginBottom: "10px",
+          textAlign: "center",
+          borderRadius: "7px",
+          letterSpacing: "16px",
+          textTransform: "uppercase",
+          color: "#5c5c5c"
         }} />
           <Header monthSelectStyles={{
           backgroundColor: "#f0f0f0",
@@ -289,6 +302,12 @@ import{D as r,a as j,j as e,T as a,H as t,W as d,b as s}from"./constants-N744Lll
           padding: "4px",
           borderRadius: "5px",
           outline: "none"
+        }} rootStyles={{
+          boxShadow: "0 2px 5px rgba(0, 0, 0, 0.1) inset",
+          borderRadius: "7px",
+          marginBottom: "10px",
+          padding: "10px",
+          backgroundColor: "#fdfdfd"
         }} />
           <WeekDays style={{
           fontWeight: 300,
@@ -308,7 +327,7 @@ import{D as r,a as j,j as e,T as a,H as t,W as d,b as s}from"./constants-N744Lll
           borderRadius: 0,
           color: "#494949"
         }} selectedStyles={{
-          backgroundColor: "#1dfc00",
+          backgroundColor: "#a4fc00",
           fontWeight: "500"
         }} weekendStyles={{
           color: "#ff0000"
@@ -320,12 +339,18 @@ import{D as r,a as j,j as e,T as a,H as t,W as d,b as s}from"./constants-N744Lll
           borderRadius: 0
         }} endOfRangeParentStyles={{
           borderRadius: 0
+        }} startOfRangeStyles={{
+          color: "#288f00"
+        }} endOfRangeStyles={{
+          color: "#288f00"
         }} inSelectedRangeParentStyles={{
-          backgroundColor: "#d5f8d5"
+          backgroundColor: "#e7ffb9"
+        }} inHoveredRangeParentStyles={{
+          backgroundColor: "#ecf8d5"
         }} inSelectedRangeStyles={{
           color: "#000000"
         }} />
         </DatePickerProvider>
       </div>;
   }
-}`,...(W=(D=f.parameters)==null?void 0:D.docs)==null?void 0:W.source}}};const B=["Example1","Example2","Example3","Example4","Example5"];export{l as Example1,i as Example2,g as Example3,c as Example4,f as Example5,B as __namedExportsOrder,T as default};
+}`,...(W=(D=f.parameters)==null?void 0:D.docs)==null?void 0:W.source}}};const E=["Example1","Example2","Example3","Example4","Example5"];export{l as Example1,i as Example2,c as Example3,g as Example4,f as Example5,E as __namedExportsOrder,T as default};